backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
waitress==3.0.0
